Get Orders of a user done

+++ src/app/modules/user/user.controller.ts
@@ -57,13 +57,26 @@ const getAllUsers = async (req: Request, res: Response) => {
 
 const getSingleUser = async (req: Request, res: Response) => {
   try {
-    const { userId } = req.params
-    const result = await UserServices.getSingleUserFromDB(userId)
-    res.status(200).json({
-      success: true,
-      message: 'A Single User is retrieved successfully!',
-      data: result,
-    })
+    const userId = parseInt(req.params.userId)
+    const userExist = await UserModel.isUserExist(userId)
+
+    if (!userExist) {
+      res.status(404).json({
+        success: false,
+        message: 'User not found',
+        error: {
+          code: 404,
+          description: 'User not found!',
+        },
+      })
+    } else {
+      const result = await UserServices.getSingleUserFromDB(userId)
+      res.status(200).json({
+        success: true,
+        message: 'A Single User is retrieved successfully!',
+        data: result,
+      })
+    }
   } catch (error: any) {
     res.status(400).json({
       success: false,
@@ -78,10 +91,8 @@ const getSingleUser = async (req: Request, res: Response) => {
 
 const updateUser = async (req: Request, res: Response) => {
   try {
-    const  userId  = parseInt(req.params.userId)
-    // check use exist or not
-    const userExist = await UserModel.isUserExist((userId))
-    // if user exists
+    const userId = parseInt(req.params.userId)
+    const userExist = await UserModel.isUserExist(userId)
     if (!userExist) {
       res.status(404).json({
         success: false,
@@ -93,7 +104,6 @@ const updateUser = async (req: Request, res: Response) => {
       })
     } else {
       const userData = req.body
-      // update user data
       const result = await UserServices.updateUser(userId, userData)
       res.status(200).json({
         success: true,
@@ -112,13 +122,26 @@ const updateUser = async (req: Request, res: Response) => {
 
 const deleteUser = async (req: Request, res: Response) => {
   try {
-    const { userId } = req.params
-    const result = await UserServices.deleteUserFromDB(userId)
-    res.status(200).json({
-      success: true,
-      message: 'User deleted successfully!',
-      data: result,
-    })
+    const userId = parseInt(req.params.userId)
+    const userExist = await UserModel.isUserExist(userId)
+
+    if (!userExist) {
+      res.status(404).json({
+        success: false,
+        message: 'User not found',
+        error: {
+          code: 404,
+          description: 'User not found!',
+        },
+      })
+    } else {
+      const result = await UserServices.deleteUserFromDB(userId)
+      res.status(200).json({
+        success: true,
+        message: 'User deleted successfully!',
+        data: result,
+      })
+    }
   } catch (error: any) {
     res.status(400).json({
       success: false,
@@ -131,10 +154,46 @@ const deleteUser = async (req: Request, res: Response) => {
   }
 }
 
+const getUserOrders = async (req: Request, res: Response): Promise<void> => {
+  try {
+    const userId = parseInt(req.params.userId)
+    // check user exist or not
+    const userExist = await UserModel.isUserExist(userId)
+
+    // if user not exist
+    if (!userExist) {
+      res.status(404).json({
+        success: false,
+        message: 'User not found',
+        error: {
+          code: 404,
+          description: 'User not found!',
+        },
+      })
+      return
+    }
+
+    // if user exist
+    const result = await UserServices.getUserOrders(userId)
+    res.status(200).json({
+      success: true,
+      message: 'Orders fetched successfully',
+      data: result,
+    })
+  } catch (error: any) {
+    res.status(500).json({
+      success: false,
+      message: error.message || 'Something went wrong',
+      error: error,
+    })
+  }
+}
+
 export const UserControllers = {
   createUser,
   getAllUsers,
   getSingleUser,
   updateUser,
   deleteUser,
+  getUserOrders,
 }
